# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Origin FGL HRMS Backend",
    "description": "Phase 2a backend contract for Approval Inbox.",
    "version": "2.0.0"
  },
  "paths": {
...
    "/api/v1/leave-requests": {
      "post": {
        "tags": ["leave_requests"],
        "summary": "Create Leave Request",
        "operationId": "create_leave_request",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LeaveRequestCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LeaveRequestResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "404": {
            "description": "Not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "409": {
            "description": "Transition conflict.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": ["leave_requests"],
        "summary": "List Leave Requests",
        "operationId": "list_leave_requests",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/LeaveStatus"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Optional status filter.",
              "title": "Status"
            },
            "description": "Optional status filter."
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Cursor from the previous page.",
              "title": "Cursor"
            },
            "description": "Cursor from the previous page."
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Page size.",
              "default": 25,
              "title": "Limit"
            },
            "description": "Page size."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LeaveRequestPage"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "404": {
            "description": "Not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "409": {
            "description": "Transition conflict.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/leave-requests/{leave_request_id}": {
      "get": {
        "tags": ["leave_requests"],
        "summary": "Get Leave Request",
        "operationId": "get_leave_request",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "leave_request_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Leave Request Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LeaveRequestResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "404": {
            "description": "Not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "409": {
            "description": "Transition conflict.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/leave-requests/{leave_request_id}/approve": {
      "post": {
        "tags": ["leave_requests"],
        "summary": "Approve Leave Request",
        "operationId": "approve_leave_request",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "leave_request_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Leave Request Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LeaveRequestResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "404": {
            "description": "Not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "409": {
            "description": "Transition conflict.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/leave-requests/{leave_request_id}/reject": {
      "post": {
        "tags": ["leave_requests"],
        "summary": "Reject Leave Request",
        "operationId": "reject_leave_request",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "leave_request_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Leave Request Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RejectLeaveRequestBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LeaveRequestResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "404": {
            "description": "Not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "409": {
            "description": "Transition conflict.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/leave-requests/{leave_request_id}/cancel": {
      "post": {
        "tags": ["leave_requests"],
        "summary": "Cancel Leave Request",
        "operationId": "cancel_leave_request",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "leave_request_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Leave Request Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LeaveRequestResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "404": {
            "description": "Not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "409": {
            "description": "Transition conflict.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "LeaveRequestCreate": {
        "properties": {
          "employee_id": {
            "type": "string",
            "title": "Employee Id",
            "description": "Requester employee id."
          },
          "type": {
            "$ref": "#/components/schemas/LeaveType",
            "description": "Leave type."
          },
          "start_date": {
            "type": "string",
            "format": "date",
            "title": "Start Date",
            "description": "Start date."
          },
          "end_date": {
            "type": "string",
            "format": "date",
            "title": "End Date",
            "description": "End date."
          },
          "note": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Note",
            "description": "Optional requester note."
          }
        },
        "type": "object",
        "required": ["employee_id", "type", "start_date", "end_date"],
        "title": "LeaveRequestCreate"
      },
      "LeaveRequestPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/LeaveRequestResponse"
            },
            "type": "array",
            "title": "Items",
            "description": "Leave request page items."
          },
          "page": {
            "$ref": "#/components/schemas/PageInfo",
            "description": "Cursor page metadata."
          }
        },
        "type": "object",
        "required": ["items", "page"],
        "title": "LeaveRequestPage"
      },
      "LeaveRequestResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "description": "Prefixed leave request id."
          },
          "employee_id": {
            "type": "string",
            "title": "Employee Id",
            "description": "Requester employee id."
          },
          "type": {
            "$ref": "#/components/schemas/LeaveType",
            "description": "Leave type."
          },
          "start_date": {
            "type": "string",
            "format": "date",
            "title": "Start Date",
            "description": "Start date."
          },
          "end_date": {
            "type": "string",
            "format": "date",
            "title": "End Date",
            "description": "End date."
          },
          "business_days": {
            "type": "integer",
            "title": "Business Days",
            "description": "Stored business-day count."
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note",
            "description": "Requester note."
          },
          "status": {
            "$ref": "#/components/schemas/LeaveStatus",
            "description": "Approval status."
          },
          "decided_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Decided By",
            "description": "Decision actor user id."
          },
          "decided_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Decided At",
            "description": "Decision timestamp."
          },
          "decision_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Decision Reason",
            "description": "Rejection reason."
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At",
            "description": "Creation timestamp."
          },
          "cancelled_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cancelled At",
            "description": "Cancellation timestamp."
          }
        },
        "type": "object",
        "required": [
          "id",
          "employee_id",
          "type",
          "start_date",
          "end_date",
          "business_days",
          "status",
          "created_at"
        ],
        "title": "LeaveRequestResponse"
      },
      "LeaveStatus": {
        "type": "string",
        "enum": ["pending", "approved", "rejected", "cancelled"],
        "title": "LeaveStatus"
      },
      "LeaveType": {
        "type": "string",
        "enum": ["ANNUAL", "SICK", "UNPAID", "EXCUSE", "MARRIAGE", "BEREAVEMENT"],
        "title": "LeaveType"
      },
...
      "PageInfo": {
        "properties": {
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor",
            "description": "Cursor for the next page, if present."
          }
        },
        "type": "object",
        "title": "PageInfo"
      },
      "ProblemDetail": {
        "properties": {
          "type": {
            "type": "string",
            "title": "Type",
            "description": "RFC 9457 problem type URI.",
            "default": "about:blank"
          },
          "title": {
            "type": "string",
            "title": "Title",
            "description": "Short problem title."
          },
          "status": {
            "type": "integer",
            "title": "Status",
            "description": "HTTP status code."
          },
          "detail": {
            "type": "string",
            "title": "Detail",
            "description": "Human-readable problem detail."
          },
          "errors": {
            "items": {
              "$ref": "#/components/schemas/ProblemFieldError"
            },
            "type": "array",
            "title": "Errors",
            "description": "Field-level errors."
          }
        },
        "type": "object",
        "required": ["title", "status", "detail"],
        "title": "ProblemDetail"
      },
      "ProblemFieldError": {
        "properties": {
          "field": {
            "type": "string",
            "title": "Field",
            "description": "Field path associated with the validation or policy error."
          },
          "code": {
            "type": "string",
            "title": "Code",
            "description": "Stable machine-readable error code."
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Human-readable English error message."
          }
        },
        "type": "object",
        "required": ["field", "code", "message"],
        "title": "ProblemFieldError"
      },
      "RejectLeaveRequestBody": {
        "properties": {
          "reason": {
            "type": "string",
            "minLength": 5,
            "title": "Reason",
            "description": "Required rejection reason."
          }
        },
        "type": "object",
        "required": ["reason"],
        "title": "RejectLeaveRequestBody"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}